+++ R/brmsframe.R
@@ -85,6 +85,14 @@ brmsframe.btl <- function(x, data, frame = list(), basis = NULL, ...) {
   # only store the ranefs of this specific linear formula
   x$frame$re <- subset2(frame$re, ls = check_prefix(x))
   class(x) <- c("bframel", class(x))
+  validate_re_s2z_design(x, data = data)
+  if (!is.null(basis[["re_s2z_center"]])) {
+    # Partial-centering fractions are part of the fitted coordinate map. Reuse
+    # them for new data rather than deriving a different map at prediction time.
+    x$sdata$re_s2z_center <- basis[["re_s2z_center"]]
+  } else {
+    x$sdata$re_s2z_center <- data_re_s2z_center(x, data = data)
+  }
   # these data_ functions may require the outputs of the corresponding
   # frame_ functions (but not vice versa) and are thus evaluated last
   x$sdata$gp <- data_gp(x, data, internal = TRUE)
@@ -348,6 +356,15 @@ frame_basis.btl <- function(x, data, ...) {
   out$sp <- frame_basis_sp(x, data, ...)
   out$ac <- frame_basis_ac(x, data, ...)
   out$bhaz <- frame_basis_bhaz(x, data, ...)
+  if (is.bframel(x) && has_re_s2z(x)) {
+    re_s2z_center <- x$sdata[["re_s2z_center"]]
+    if (is.null(re_s2z_center)) {
+      re_s2z_center <- data_re_s2z_center(x, data = data)
+    }
+    if (length(re_s2z_center)) {
+      out$re_s2z_center <- re_s2z_center
+    }
+  }
   out
 }